Make installation happen in a temporary directory, remove package-lock usage
