+++ cli/api/BUILD
@@ -22,19 +22,23 @@ ts_library(
         "//sqlx:lexer",
         "//cli/vm:vm",
         "@npm//@google-cloud/bigquery",
+        "@npm//@types/fs-extra",
         "@npm//@types/glob",
         "@npm//@types/js-beautify",
         "@npm//@types/js-yaml",
         "@npm//@types/long",
         "@npm//@types/node",
         "@npm//@types/semver",
+        "@npm//@types/tmp",
         "@npm//deepmerge",
+        "@npm//fs-extra",
         "@npm//glob",
         "@npm//js-beautify",
         "@npm//js-yaml",
         "@npm//promise-pool-executor",
         "@npm//protobufjs",
         "@npm//semver",
+        "@npm//tmp",
         "@npm//sql-formatter",
     ],
 )
