+++ package.json
@@ -42,6 +42,7 @@
     "@types/semver": "^7.3.13",
     "@types/stack-trace": "^0.0.29",
     "@types/swagger-schema-official": "^2.0.20",
+    "@types/tmp": "^0.2.0",
     "@types/uuid": "^7.0.2",
     "@types/validator": "^13.1.0",
     "@types/vscode": "^1.45.1",
@@ -108,7 +109,7 @@
     "sql-formatter": "^12.2.1",
     "style-loader": "^1.0.0",
     "tarjan-graph": "^2.0.0",
-    "tmp": "0.2.0",
+    "tmp": "^0.2.0",
     "ts-loader": "^5.3.1",
     "ts-mockito": "^2.6.1",
     "tslint": "^5.14.0",
