+++ packages/@dataform/cli/BUILD
@@ -27,26 +27,27 @@ nodejs_binary(
 externals = [
     "@google-cloud/bigquery",
     "@segment/analytics-node",
-    "protobufjs",
-    "tarjan-graph",
-    "semver",
-    "moo",
     "chokidar",
-    "uuid",
-    "readline-sync",
-    "untildify",
-    "yargs",
-    "glob",
     "deepmerge",
+    "fs-extra",
     "glob",
     "js-beautify",
     "js-yaml",
+    "moo",
     "object-sizeof",
     "parse-duration",
     "promise-pool-executor",
+    "protobufjs",
+    "readline-sync",
+    "semver",
     "sql-formatter",
+    "tarjan-graph",
+    "tmp",
     "typeid-js",
+    "uuid",
+    "untildify",
     "vm2",
+    "yargs",
 ]
 
 pkg_json(
